@@ -43,4 +43,4 @@ class IncidentPriority(Names):
         return f"{self.__class__.__name__}[{self.name!r}]"
 
     def __str__(self) -> str:
-        return self.name
+        return self.name  # type: ignore[no-any-return]
